# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Todo API",
    "description": "The API exposes a Todo list",
    "version": "v1"
  },
  "paths": {
    "/api/TodoItems": {
      "get": {
        "tags": [
          "TodoItems"
        ],
        "summary": "Get the complete todo list",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TodoItems"
        ],
        "summary": "Creates a TodoItem.",
        "description": "Sample request:\r\n            \r\n    POST /Todo\r\n    {\r\n       \"id\": 100,\r\n       \"text\": \"a task that has to be done\",\r\n       \"completed\": false\r\n    }",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              }
            }
          },
          "400": {
            "description": "If there is already an item with the same id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Returns the newly created item"
          }
        }
      }
    },
    "/api/TodoItems/{id}": {
      "get": {
        "tags": [
          "TodoItems"
        ],
        "summary": "Get a specific TodoItem.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier of the requested Todo Item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "TodoItem": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "completed": {
            "type": "boolean",
            "description": "Indicates if the todo item was completed",
            "default": false
          },
          "text": {
            "pattern": "^[a-zA-Z''-'\\s]{1,40}$",
            "type": "string",
            "description": "The todo item description"
          }
        },
        "additionalProperties": false
      }
    }
  }
}